+++ lib/alipay.rb
@@ -14,15 +14,14 @@ require 'alipay/wap/notify'
 require 'alipay/wap/sign'
 require 'alipay/mobile/service'
 require 'alipay/mobile/sign'
-require 'alipay/app/service'
-require 'alipay/app/sign'
+require 'alipay/client'
 
 module Alipay
   @debug_mode = true
   @sign_type = 'MD5'
 
   class << self
-    attr_accessor :pid, :key, :sign_type, :debug_mode
+    attr_accessor :pid, :key, :legacy_gateway_url, :sign_type, :debug_mode
 
     def debug_mode?
       !!@debug_mode