+++ test/alipay/notify_test.rb
@@ -14,21 +14,21 @@ class Alipay::NotifyTest < Minitest::Test
 
   def test_unsign_notify
     stub_request(
-      :get, "https://mapi.alipay.com/gateway.do?service=notify_verify&partner=#{Alipay.pid}&notify_id=1234"
+      :get, "#{Alipay.legacy_gateway_url}/gateway.do?service=notify_verify&partner=#{Alipay.pid}&notify_id=1234"
     ).to_return(body: "true")
     assert !Alipay::Notify.verify?(@unsign_params)
   end
 
   def test_verify_notify_when_true
     stub_request(
-      :get, "https://mapi.alipay.com/gateway.do?service=notify_verify&partner=#{Alipay.pid}&notify_id=1234"
+      :get, "#{Alipay.legacy_gateway_url}/gateway.do?service=notify_verify&partner=#{Alipay.pid}&notify_id=1234"
     ).to_return(body: "true")
     assert Alipay::Notify.verify?(@sign_params)
   end
 
   def test_verify_notify_when_false
     stub_request(
-      :get, "https://mapi.alipay.com/gateway.do?service=notify_verify&partner=#{Alipay.pid}&notify_id=1234"
+      :get, "#{Alipay.legacy_gateway_url}/gateway.do?service=notify_verify&partner=#{Alipay.pid}&notify_id=1234"
     ).to_return(body: "false")
     assert !Alipay::Notify.verify?(@sign_params)
   end